% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_attribues.R
\name{compare_attributes}
\alias{compare_attributes}
\title{Compare attributes}
\usage{
compare_attributes(x, y)
}
\arguments{
\item{x, y}{Objects for which to compare attributes.}
}
\description{
Compares attributes of two objects.
}
