% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_project.R
\name{copy_directory}
\alias{copy_directory}
\title{Recursively copy a directory.}
\usage{
copy_directory(path, new_path, all = TRUE)
}
\arguments{
\item{path}{Path of the directory to copy.}

\item{new_path}{Path to copy directory to.}

\item{all}{Boolean. If \code{TRUE}, will copy hidden files as well.}
}
\value{
Prints out messages saying which sub-directories it created.
}
\description{
Recursively copies a directory, printing out messages saying which
sub-directories were created.
}
\examples{
\donttest{copy_directory(path = "old/dir", new_path = "new/dir")}
}
