% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_project.R
\name{copy_file}
\alias{copy_file}
\title{Copies files and recursively copies directories.}
\usage{
copy_file(path, new_path, all = TRUE)
}
\arguments{
\item{path}{Path of the file to copy.}

\item{new_path}{Path to copy file to.}

\item{all}{Boolean. If \code{TRUE}, will copy hidden files as well.}
}
\value{
Prints out messages saying which sub-directories it created.
}
\description{
Recursively copies a file, printing out messages saying which
sub-directories were created.
}
\examples{
\donttest{copy_file(path = "old/dir", new_path = "new/dir")}
}
