% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_project.R
\name{create_data_project}
\alias{create_data_project}
\title{Create project folder}
\usage{
create_data_project(
  path,
  template_url = "https://github.com/dcl-docs/project/archive/master.zip"
)
}
\arguments{
\item{path}{Path for project folder.}

\item{template_url}{Download URL of template repository}
}
\description{
Create a project folder with Makefile and sub-folders: data-raw, data,
scripts, docs, eda, and reports.
Available from RStudio: File > New Project... > New Directory > DCL project
}
