% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_project.R
\name{create_directory_from_template}
\alias{create_directory_from_template}
\title{Create and populate a directory from a template}
\usage{
create_directory_from_template(path, url_template)
}
\arguments{
\item{path}{Path to create directory.}

\item{url_template}{A GitHub repository download link.}
}
\value{
Prints out messages saying the directory were created.
}
\description{
Recursively creates a directory using a template downloaded from
the internet.
}
\examples{
\dontrun{
create_directory_from_template(
  "new_dir",
  "https://github.com/dcl-docs/project/archive/master.zip"
)
}
}
\keyword{internal}
