% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_project.R
\name{dir_empty}
\alias{dir_empty}
\title{Check if a directory is empty}
\usage{
dir_empty(path)
}
\arguments{
\item{path}{Path to directory}
}
\value{
Boolean.
}
\description{
Check if a directory is empty
}
\keyword{internal}
