% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_download_unzip.R
\name{file_download_unzip}
\alias{file_download_unzip}
\title{Download and unzip a file.}
\usage{
file_download_unzip(url_file, path)
}
\arguments{
\item{url_file}{URL of the file to download and unzip.}

\item{path}{Directory to put the unzipped file.}
}
\value{
Character vector giving the path to the downloaded and unzipped file.
}
\description{
Downloads a zipped file from the internet and unzips it to a directory.
}
\examples{
\donttest{
file_download_unzip(
  url_file = "https://github.com/dcl-docs/project/archive/master.zip",
  path = "~/my/dir"
)
}
}
\keyword{internal}
