% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Create project folder}
\usage{
make_project(path, ...)
}
\arguments{
\item{path}{Path for project folder.}

\item{...}{Ignored.}
}
\description{
Create a project folder with Makefile and sub-folders: data-raw, data,
scripts, docs, eda, and reports.
Available from RStudio: File > New Project... > New Directory > DCL project
}
