% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare two objects}
\usage{
compare(model, comparison, ignore_attr = NULL, ...)
}
\arguments{
\item{model}{1}

\item{comparison}{2}

\item{...}{3}
}
\description{
Compare two objects using \link[compare:compare]{compare::compare}, ignoring specified attributes.
}
\examples{
# Ignore spec attribute created by readr functions
compare(answers$q1, q1, ignore_attr = "spec")

}
