% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare two objects}
\usage{
compare(model, comparison, ignore_attrs = NULL, ...)
}
\arguments{
\item{model}{The "correct" object.}

\item{comparison}{The object to be compared with \code{model}.}

\item{ignore_attrs}{String or vector of strings specifying attributes to ignore.}

\item{...}{Arguments to be passed to \code{compare::compare()}.}
}
\description{
Compare two objects using \code{\link[compare:compare]{compare::compare()}}, ignoring attributes specified
with \code{ignore_attrs}.
}
\examples{
# Ignore `spec` attribute
compare(answers$q1, q1, ignore_attrs = "spec")

# Ignore `spec` and `problems` attributes
compare(answers$q1, q1, ignore_attrs = c("spec", "problems"))

}
