% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare two objects}
\usage{
compare(model, comparison, compare_attrs = c("class", "names", "groups",
  "row.names"), ignore_classes = "spec_tbl_df",
  ignore_component_attrs = TRUE, ...)
}
\arguments{
\item{model}{The "correct" object.}

\item{comparison}{The object to be compared with \code{model}.}

\item{compare_attrs}{String or vector of strings of attributes to compare.
Default is to compare only \code{class}, \code{groups}, \code{names}, and \code{row.names}
attributes. Use \code{NULL} to compare all attributes.}

\item{ignore_classes}{String or vector of strings of classes to ignore.
Default is to ignore \code{spec_tbl_df} class. Use \code{NULL} to not ignore classes.}

\item{ignore_component_attrs}{Ignore attributes of object components.}

\item{...}{Arguments to be passed to \code{compare::compare()}.}
}
\description{
Compare two objects using \code{\link[compare:compare]{compare::compare()}}, comparing only the attributes
in \code{compare_attrs} and ignoring classes in \code{ignore_classes}.
}
\examples{
df1 <- data.frame(x = 1, y = 2)
df2 <- df1
attr(df2, "class") <- c("spec_tbl_df", attr(df1, "class"))
attr(df2, "foo") <- "bar"

# Compare objects
compare(df1, df2)

}
