% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare two objects}
\usage{
compare(model, comparison, ...)
}
\arguments{
\item{model}{!!!!!}

\item{...}{}
}
\description{
This uses \link[compare:compare]{compare::compare} with some tweaks designed to
checking of DCL exercises easier.
}
\examples{
# This is a comment
compare(mtcars, mtcars)
}
