% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare two objects}
\usage{
compare(model, comparison, ignore_attrs = "spec",
  ignore_classes = "spec_tbl_df", ...)
}
\arguments{
\item{model}{The "correct" object.}

\item{comparison}{The object to be compared with \code{model}.}

\item{ignore_attrs}{String or vector of strings of attributes to ignore.
Default is to ignore \code{spec} attribute. Use \code{NULL} to not ignore attributes.}

\item{ignore_classes}{String or vector of strings of classes to ignore.
Default is to ignore \code{spec_tbl_df} class. Use \code{NULL} to not ignore classes.}

\item{...}{Arguments to be passed to \code{compare::compare()}.}
}
\description{
Compare two objects using \code{\link[compare:compare]{compare::compare()}}, ignoring \code{spec} attribute and
\code{spec_tbl_df} class by default.
}
\examples{
df1 <- data.frame(x = 1, y = 2)
df2 <- df1
attr(df2, "spec") <- "a"
attr(df2, "class") <- c("spec_tbl_df", attr(df1, "class"))
df3 <- df2
attr(df3, "problems") <- "b"

# Ignore spec attribute and spec_tbl_df class by default
compare(df1, df2)

# Do not ignore attributes or classes
compare(df1, df2, ignore_attrs = NULL, ignore_classes = NULL)

# Ignore problems attribute
compare(df2, df3, ignore_attrs = "problems")

# Ignore spec and problems attributes
compare(df1, df3, ignore_attrs = c("spec", "problems"))

}
