% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_project.R
\name{create_data_project}
\alias{create_data_project}
\title{Create a project from a template}
\usage{
create_data_project(
  path,
  project = TRUE,
  template = "https://github.com/dcl-docs/project/archive/master.zip",
  open = interactive()
)
}
\arguments{
\item{path}{A path giving the location for the new project. If it does not already exist, it is created.}

\item{project}{If \code{TRUE}, creates an RStudio project.}

\item{template}{Download URL of template repository. Defaults to the dcl-docs
project template.}

\item{open}{If \code{TRUE} and \code{project} is \code{TRUE}, activates the project. If you're
using RStudio, this will result in the project opening in a new session.}
}
\description{
Creates a project from a template. By default, creates a project
from this template: \url{https://github.com/dcl-docs/project}
}
\examples{
\dontrun{create_data_project("path/to/project")}
}
