% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_project.R
\name{create_data_project}
\alias{create_data_project}
\title{Create a project from a template}
\usage{
create_data_project(
  path,
  project = TRUE,
  template = "https://github.com/dcl-docs/project/archive/master.zip",
  rstudio = rstudioapi::isAvailable(),
  open = interactive()
)
}
\arguments{
\item{path}{Path for project folder.}

\item{project}{Should this be an RStudio project?}

\item{template}{Download URL of template repository}

\item{rstudio}{Are you using RStudio?}

\item{open}{Boolean. If \code{project} is \code{TRUE} and \code{open} is \code{TRUE}, will
open the project in a new session.}
}
\description{
Creates a project from a zipped template file. By default, creates a project
with 6 folders:
}
\examples{
\donttest{create_data_project("path/to/project")}
}
